Module std_detect::detect::arch::x86::__is_feature_detected

🔬This is a nightly-only experimental API. (stdsimd)

Each function performs run-time feature detection for a single feature. This allow us to use stability attributes on a per feature basis.

PLEASE: do not use this, it is an implementation detail subject to change.

Functions

adx

PLEASE: do not use this, it is an implementation detail subject to change.

aes

PLEASE: do not use this, it is an implementation detail subject to change.

avx

PLEASE: do not use this, it is an implementation detail subject to change.

avx2

PLEASE: do not use this, it is an implementation detail subject to change.

avx512bf16

PLEASE: do not use this, it is an implementation detail subject to change.

avx512bitalg

PLEASE: do not use this, it is an implementation detail subject to change.

avx512bw

PLEASE: do not use this, it is an implementation detail subject to change.

avx512cd

PLEASE: do not use this, it is an implementation detail subject to change.

avx512dq

PLEASE: do not use this, it is an implementation detail subject to change.

avx512er

PLEASE: do not use this, it is an implementation detail subject to change.

avx512f

PLEASE: do not use this, it is an implementation detail subject to change.

avx512ifma

PLEASE: do not use this, it is an implementation detail subject to change.

avx512pf

PLEASE: do not use this, it is an implementation detail subject to change.

avx512vbmi

PLEASE: do not use this, it is an implementation detail subject to change.

avx512vbmi2

PLEASE: do not use this, it is an implementation detail subject to change.

avx512vl

PLEASE: do not use this, it is an implementation detail subject to change.

avx512vnni

PLEASE: do not use this, it is an implementation detail subject to change.

avx512vp2intersect

PLEASE: do not use this, it is an implementation detail subject to change.

avx512vpopcntdq

PLEASE: do not use this, it is an implementation detail subject to change.

bmi1

PLEASE: do not use this, it is an implementation detail subject to change.

bmi2

PLEASE: do not use this, it is an implementation detail subject to change.

cmpxchg16b

PLEASE: do not use this, it is an implementation detail subject to change.

ermsb

PLEASE: do not use this, it is an implementation detail subject to change.

f16c

PLEASE: do not use this, it is an implementation detail subject to change.

fma

PLEASE: do not use this, it is an implementation detail subject to change.

fxsr

PLEASE: do not use this, it is an implementation detail subject to change.

gfni

PLEASE: do not use this, it is an implementation detail subject to change.

lzcnt

PLEASE: do not use this, it is an implementation detail subject to change.

mmx

PLEASE: do not use this, it is an implementation detail subject to change.

movbe

PLEASE: do not use this, it is an implementation detail subject to change.

pclmulqdq

PLEASE: do not use this, it is an implementation detail subject to change.

popcnt

PLEASE: do not use this, it is an implementation detail subject to change.

rdrand

PLEASE: do not use this, it is an implementation detail subject to change.

rdseed

PLEASE: do not use this, it is an implementation detail subject to change.

rtm

PLEASE: do not use this, it is an implementation detail subject to change.

sha

PLEASE: do not use this, it is an implementation detail subject to change.

sse

PLEASE: do not use this, it is an implementation detail subject to change.

sse2

PLEASE: do not use this, it is an implementation detail subject to change.

sse3

PLEASE: do not use this, it is an implementation detail subject to change.

sse4_1

PLEASE: do not use this, it is an implementation detail subject to change.

sse4_2

PLEASE: do not use this, it is an implementation detail subject to change.

sse4a

PLEASE: do not use this, it is an implementation detail subject to change.

ssse3

PLEASE: do not use this, it is an implementation detail subject to change.

tbm

PLEASE: do not use this, it is an implementation detail subject to change.

tsc

PLEASE: do not use this, it is an implementation detail subject to change.

vaes

PLEASE: do not use this, it is an implementation detail subject to change.

vpclmulqdq

PLEASE: do not use this, it is an implementation detail subject to change.

xsave

PLEASE: do not use this, it is an implementation detail subject to change.

xsavec

PLEASE: do not use this, it is an implementation detail subject to change.

xsaveopt

PLEASE: do not use this, it is an implementation detail subject to change.

xsaves

PLEASE: do not use this, it is an implementation detail subject to change.