Crate std_detect

🔬This is a nightly-only experimental API. (stdsimd)

Run-time feature detection for the Rust standard library.

To detect whether a feature is enabled in the system running the binary use one of the appropriate macro for the target:

x86 and x86_64: is_x86_feature_detected

arm: is_arm_feature_detected

aarch64: is_aarch64_feature_detected

riscv: is_riscv_feature_detected

mips: is_mips_feature_detected

mips64: is_mips64_feature_detected

powerpc: is_powerpc_feature_detected

powerpc64: is_powerpc64_feature_detected

Modules

detectExperimental

This module implements run-time feature detection.

Macros

detect_featureExperimental

is_arm_feature_detectedExperimentalARM

Checks if arm feature is enabled.

is_mips64_feature_detectedExperimentalMIPS-64

Checks if mips64 feature is enabled.

is_mips_feature_detectedExperimentalMIPS

Checks if mips feature is enabled.

is_powerpc64_feature_detectedExperimentalPowerPC-64

Checks if powerpc feature is enabled.

is_powerpc_feature_detectedExperimentalPowerPC

Checks if powerpc feature is enabled.

is_riscv_feature_detectedExperimentalRISC-V RV32 or RISC-V RV64

A macro to test at runtime whether instruction sets are available on RISC-V platforms.

is_aarch64_feature_detectedAArch64

This macro tests, at runtime, whether an aarch64 feature is enabled on aarch64 platforms. Currently most features are only supported on linux-based platforms.

is_x86_feature_detectedx86 or x86-64

A macro to test at runtime whether a CPU feature is available on x86/x86-64 platforms.